% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allow_thumbnails.R
\name{allow_thumbnails}
\alias{allow_thumbnails}
\title{Generate a full-resolution pdf image before generating a smaller image}
\usage{
allow_thumbnails(x, options)
}
\arguments{
\item{x}{A pdf image.}

\item{options}{Additional settings.}
}
\value{
A full resolution pdf image.
}
\description{
Generate a full-resolution pdf image before generating a smaller image
}
