% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arriba.R
\name{arriba_pdf_read}
\alias{arriba_pdf_read}
\title{Read Arriba PDF File}
\usage{
arriba_pdf_read(pdf = NULL, fusions = NULL, outdir = NULL)
}
\arguments{
\item{pdf}{Output PDF file from Arriba.}

\item{fusions}{Tibble with fusions from Arriba.}

\item{outdir}{Directory to write output PNGs to.}
}
\value{
Tibble with paths to the created PNG images and their clean names.
If the PDF (or any of the required params) is NULL, returns NULL.
|> |> |> |>  @examples
pdf <- system.file("rawdata/test_data/dragen/arriba/fusions.pdf", package = "RNAsum")
tsv <- system.file("rawdata/test_data/dragen/arriba/fusions.tsv", package = "RNAsum")
fusions <- arriba_tsv_read(tsv)
(pngs <- arriba_pdf_read(pdf, fusions, tempdir()))
}
\description{
Reads the PDF output by Arriba containing one fusion plot per page and
converts each page to a separate PNG file.
}
