% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arriba_plots.R
\name{arriba_plots}
\alias{arriba_plots}
\title{Fusion visualisation}
\usage{
arriba_plots(arriba_file, arriba_results, results_dir)
}
\arguments{
\item{arriba_file}{Output fusions from Arriba.}

\item{arriba_results}{PDF files.}

\item{results_dir}{Results directory.}
}
\value{
PNG images.
}
\description{
Fusion output visualisation.
}
