% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arriba.R
\name{arriba_process}
\alias{arriba_process}
\title{Process Arriba Fusions}
\usage{
arriba_process(arriba.fusions, known_translocations, genes_cancer)
}
\arguments{
\item{arriba.fusions}{Parsed tibble with Arriba fusions.}

\item{known_translocations}{Tibble with known translocations.}

\item{genes_cancer}{Character vector of cancer genes.}
}
\value{
Tibble with post-processed Arriba fusion calls.
}
\description{
Process Arriba Fusions
}
