% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arriba.R
\name{arriba_read_pdf}
\alias{arriba_read_pdf}
\title{Read Arriba PDF File}
\usage{
arriba_read_pdf(pdf = NULL, fusions = NULL, outdir = NULL)
}
\arguments{
\item{pdf}{Output PDF file from Arriba.}

\item{fusions}{Tibble with fusions from Arriba.}

\item{outdir}{Directory to write output PNGs to.}
}
\value{
Tibble with paths to the created PNG images and their clean names.
If the PDF (or any of the required params) is NULL, returns NULL.
}
\description{
Reads the PDF output by Arriba containing one fusion plot per page and
converts each page to a separate PNG file.
}
\examples{
pdf <- system.file("rawdata/test_data/dragen/arriba/fusions.pdf", package = "RNAsum")
tsv <- system.file("rawdata/test_data/dragen/arriba/fusions.tsv", package = "RNAsum")
fusions <- arriba_read_tsv(tsv)
(pngs <- arriba_read_pdf(pdf, fusions, tempdir()))
}
