% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arriba_plots.R
\name{arriba_read_tsv}
\alias{arriba_read_tsv}
\title{Read Arriba TSV File}
\usage{
arriba_read_tsv(x)
}
\arguments{
\item{x}{Path to Arriba TSV file.}
}
\value{
A tibble with the contents of the input TSV file.
}
\description{
Reads the TSV file output by Arriba.
}
\examples{
x <- system.file("rawdata/test_data/dragen/arriba/fusions.tsv", package = "RNAsum")
(a <- arriba_read_tsv(x))
}
