% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions_arriba.R
\name{arriba_summary_write}
\alias{arriba_summary_write}
\title{Write Arriba Fusion Summary}
\usage{
arriba_summary_write(x, file)
}
\arguments{
\item{x}{Tibble with Arriba fusions or NULL.}

\item{file}{File to write the summary info to.}
}
\value{
The input invisibly.
}
\description{
Writes Arriba fusion summary information (\code{gene1__gene2_bp1_bp2}).
}
