% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arriba_plots.R
\name{arriba_tsv_read}
\alias{arriba_tsv_read}
\title{Read Arriba TSV File}
\usage{
arriba_tsv_read(x)
}
\arguments{
\item{x}{Path to arriba TSV file.}
}
\value{
A tibble with the contents of the input TSV file.
}
\description{
Reads the TSV file output by arriba.
}
\examples{
x <- system.file("rawdata/test_data/dragen/arriba/fusions.tsv", package = "RNAsum")
(a <- arriba_tsv_read(x))
}
