% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{barPlot}
\alias{barPlot}
\title{Generate bar-plot for selected genes}
\usage{
barPlot(
  gene,
  data,
  targets,
  y_title = "Counts",
  sampleName,
  ext_cancer,
  int_cancer,
  add_cancer = NULL
)
}
\arguments{
\item{gene}{User defined gene.}

\item{data}{Input data.}

\item{targets}{Targets.}

\item{y_title}{Title for y-axis.}

\item{sampleName}{Sample name.}

\item{ext_cancer}{External cancer group.}

\item{int_cancer}{Internal cancer group.}

\item{add_cancer}{Used for reordering groups.}
}
\value{
Bar-plot for selected genes, highlighting samples of interest.
}
\description{
Generates bar-plot for selected genes, highlighting samples of interest.
}
