% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfPlot.R
\name{cdfPlot}
\alias{cdfPlot}
\title{Generates cumulative distribution function (CDF) plot for selected gene.}
\usage{
cdfPlot(
  gene,
  data,
  targets,
  sampleName,
  int_cancer,
  ext_cancer,
  comp_cancer,
  add_cancer = NULL,
  addBoxPlot = FALSE,
  scaling = "gene-wise",
  report_dir
)
}
\arguments{
\item{gene}{Gene of interest.}

\item{data}{Input data.}

\item{targets}{Target data set.}

\item{sampleName}{Sample name.}

\item{int_cancer}{Internal cancer group.}

\item{ext_cancer}{External cancer group.}

\item{comp_cancer}{Complete cancer group.}

\item{add_cancer}{Used for reordering groups.}

\item{addBoxPlot}{Add boc plot (boolean).}

\item{scaling}{Gene-wise or sample-wise.}

\item{report_dir}{Output directory.}
}
\value{
Cumulative distribution function (CDF) plot for selected gene.
}
\description{
Generates cumulative distribution function (CDF) plot for selected gene.
}
