% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicDrugTable.R
\name{civicDrugTable}
\alias{civicDrugTable}
\title{Generate table with drugs targeting selected set of genes}
\usage{
civicDrugTable(
  genes,
  civic_var_summaries,
  civic_clin_evid,
  evid_type = "Predictive",
  var_type = NULL
)
}
\arguments{
\item{genes}{Genes of interest.}

\item{civic_var_summaries}{CIViC variant summaries.}

\item{civic_clin_evid}{CIViC clinical evidence.}

\item{evid_type}{Evidence type.}

\item{var_type}{Variant type.}
}
\value{
Table with drugs targeting selected set of genes.
}
\description{
Generates table with drugs targeting selected set of genes using info from
CIViC database (https://civicdb.org/).
}
