% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_info.R
\name{clinical_info_read}
\alias{clinical_info_read}
\title{Read Clinical Information Spreadsheet}
\usage{
clinical_info_read(x = NULL)
}
\arguments{
\item{x}{Path to spreadsheet containing clinical information.}
}
\value{
A data.frame representation of the input spreadsheet, or NULL if input
is NULL.
}
\description{
Reads clinical information spreadsheet.
}
\examples{
x <- system.file("rawdata/test_data/test_clinical_data.xlsx", package = "RNAsum")
(d <- clinical_info_read(x))
}
