% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgr.R
\name{cn_subset}
\alias{cn_subset}
\title{Copy Number Subset Table}
\usage{
cn_subset(gene.mut = NULL, cn_data, expr_data.z, expr_data.perc)
}
\arguments{
\item{gene.mut}{Tibble with gene and alterations, if PCGR was run.}

\item{cn_data}{Tibble with gene and MeanCopyNumber}

\item{expr_data.z}{Tibble with SYMBOL and z-score expr values.}

\item{expr_data.perc}{Tibble with SYMBOL and percentile expr values.}
}
\value{
A tibble with Gene, CN, Diff_Perc, Diff_Z_score, and Alterations if
PCGR was run.
}
\description{
Copy Number Subset Table
}
