% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_ordered.R
\name{cumsum_ordered}
\alias{cumsum_ordered}
\title{Calculate cumulative sum}
\usage{
cumsum_ordered(x)
}
\arguments{
\item{x}{Input}
}
\value{
Cumulative sum while keeping the order of input data
}
\description{
Calculate cumulative sum
}
