% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{density2freq}
\alias{density2freq}
\title{Title Generate density and expression distribution plots for selected gene, highlighting samples of interest.}
\usage{
density2freq(density)
}
\arguments{
\item{gene}{Selected gene.}

\item{data}{Input data.}

\item{main_title}{Main title.}

\item{x_title}{X-axis title.}

\item{sampleName}{Sample name.}

\item{distributions}{Distributions of interest.}

\item{scaling}{Scaling.}
}
\value{
Density and expression distribution plots for selected gene.
}
\description{
Title Generate density and expression distribution plots for selected gene, highlighting samples of interest.
}
