% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Generate density and expression distribution plots for selected gene}
\usage{
densityPlot(
  gene,
  data,
  main_title,
  x_title,
  sampleName,
  distributions = NULL,
  scaling = "gene-wise"
)
}
\arguments{
\item{gene}{Selected gene.}

\item{data}{Input data.}

\item{main_title}{Main title.}

\item{x_title}{X-axis title.}

\item{sampleName}{Sample name.}

\item{distributions}{Distributions of interest.}

\item{scaling}{Scaling.}
}
\value{
Density and expression distribution plots for selected gene.
}
\description{
Generate density and expression distribution plots for selected gene, highlighting samples of interest.
}
