% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions_dragen.R
\name{dragen_fusions_process}
\alias{dragen_fusions_process}
\title{Process DRAGEN Fusions}
\usage{
dragen_fusions_process(dragen.fusions, known_translocations, genes_cancer)
}
\arguments{
\item{dragen.fusions}{Parsed tibble with DRAGEN fusions.}

\item{known_translocations}{Tibble with known translocations.}

\item{genes_cancer}{Character vector of cancer genes.}
}
\value{
Tibble with post-processed DRAGEN fusion calls.
}
\description{
Process DRAGEN Fusions
}
