% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_dragen.R
\name{dragen_mapping_metrics_read}
\alias{dragen_mapping_metrics_read}
\title{Read DRAGEN Mapping Metrics}
\usage{
dragen_mapping_metrics_read(x = NULL)
}
\arguments{
\item{x}{Path to the \code{mapping_metrics.csv} file.}
}
\value{
A tibble with the contents of the input TSV file, or NULL if input is NULL or TSV has no data rows.
}
\description{
Reads the \code{mapping_metrics.csv} file output by DRAGEN.
}
\examples{
x <- system.file("rawdata/test_data/dragen",
  "test.mapping_metrics.csv",
  package = "RNAsum"
)
(d1 <- dragen_mapping_metrics_read(x))
}
