% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_tbl}
\alias{empty_tbl}
\title{Create Empty Tibble Given Column Names}
\usage{
empty_tbl(cnames, ctypes = readr::cols(.default = "c"))
}
\arguments{
\item{cnames}{Character vector of column names to use.}
}
\value{
A tibble with 0 rows and the given column names.
}
\description{
From https://stackoverflow.com/a/62535671/2169986. Useful for handling
edge cases with empty data.
}
