% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprGroupsStats_geneWise.R
\name{exprGroupsStats_geneWise}
\alias{exprGroupsStats_geneWise}
\title{Calculate gene-wise median, sd, quantiles and cumulative fractions for expression data}
\usage{
exprGroupsStats_geneWise(data, targets)
}
\arguments{
\item{data}{input data.}

\item{targets}{Target groups}
}
\value{
Gene-wise median, standard deviation, quantiles and cumulative fractions for expression data.
}
\description{
Calculates gene-wise median, sd, quantiles and cumulative fractions for expression data.
}
