% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprGroupStats_groupWise.R
\name{exprGroupStats_groupWise}
\alias{exprGroupStats_groupWise}
\title{Calculate group-wise median, standard deviation, quantiles and cumulative fractions.}
\usage{
exprGroupStats_groupWise(data, targets, target)
}
\arguments{
\item{data}{Input data.}

\item{targets}{Target groups.}

\item{target}{Defined biological group.}
}
\value{
Group-wise median, standard deviation, quantiles and cumulative fractions for expression data from specific sample group.
}
\description{
Calculates group-wise median, sd, quantiles and cumulative
fractions for expression data from specific sample group.
}
