% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprTable.R
\name{exprTable}
\alias{exprTable}
\title{Generate table with coloured cells indicating expression values for selected genes}
\usage{
exprTable(
  genes,
  keep_all = FALSE,
  data,
  cn_data = NULL,
  sv_data = NULL,
  cn_decrease = TRUE,
  targets,
  sampleName,
  int_cancer,
  ext_cancer,
  comp_cancer,
  add_cancer = NULL,
  genes_annot = NULL,
  oncokb_annot = NULL,
  cancer_genes = NULL,
  mut_annot = NULL,
  fusion_genes = NULL,
  ext_links = FALSE,
  type = "z",
  scaling = "gene-wise",
  caner_genes_annot.list
)
}
\arguments{
\item{genes}{Selected genes.}

\item{keep_all}{keep all rows}

\item{data}{Input data.}

\item{cn_data}{Copy number data.}

\item{sv_data}{SV data.}

\item{cn_decrease}{Order of the CN data.}

\item{targets}{Target groups.}

\item{sampleName}{Sample name.}

\item{int_cancer}{Internal cancer group.}

\item{ext_cancer}{External cancer group.}

\item{comp_cancer}{Complete cancer group.}

\item{add_cancer}{Additonal cancer type.}

\item{genes_annot}{Geners annotation.}

\item{oncokb_annot}{OncoKb annotation.}

\item{cancer_genes}{Cancer genes.}

\item{mut_annot}{Mutation annotation.}

\item{fusion_genes}{Fusion genes.}

\item{ext_links}{External links.}

\item{type}{Type.}

\item{scaling}{Scaling}

\item{caner_genes_annot.list}{The caner_genes_annot.list from the report.}
}
\value{
Table with coloured cells indicating expression values for selected genes
}
\description{
Generates table with coloured cells indicating expression values for selected genes.
}
