% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions.R
\name{fusions_annot}
\alias{fusions_annot}
\title{Annotate Fusions with Ensembl Coordinates}
\usage{
fusions_annot(fusions, gene_ann)
}
\arguments{
\item{fusions}{Tibble with DRAGEN and Arriba fusions.}

\item{gene_ann}{Dataframe with gene annotations.}
}
\value{
Tibble with annotated genes.
}
\description{
Annotate Fusions with Ensembl Coordinates
}
