% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions.R
\name{fusions_preprocess}
\alias{fusions_preprocess}
\title{Pre-Process Arriba or DRAGEN Fusions}
\usage{
fusions_preprocess(d, known_translocations, genes_cancer)
}
\arguments{
\item{d}{Parsed tibble with Arriba or DRAGEN fusions.}

\item{known_translocations}{Tibble with known translocations.}

\item{genes_cancer}{Character vector of cancer genes.}
}
\value{
Tibble with post-processed fusion calls.
}
\description{
Pre-Process Arriba or DRAGEN Fusions
}
