% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{fusions_summary}
\alias{fusions_summary}
\title{Get Arriba/DRAGEN Fusions Summary}
\usage{
fusions_summary(tbl = NULL)
}
\arguments{
\item{tbl}{Tibble with fusions in columns 'gene1' and 'gene2'.}
}
\value{
Character vector of the 'gene1' and 'gene2' genes.
}
\description{
Get Arriba/DRAGEN Fusions Summary
}
