% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions.R
\name{fusions_table}
\alias{fusions_table}
\title{Fusions Table Display}
\usage{
fusions_table(fusions)
}
\arguments{
\item{fusions}{Tibble with fusions.}
}
\value{
A DT object.
}
\description{
Fusions Table Display
}
