% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{genes_pcgr_summary}
\alias{genes_pcgr_summary}
\title{Get Mutated Genes from PCGR}
\usage{
genes_pcgr_summary(pcgr_tbl, tiers = c(1:4), splice_vars = TRUE)
}
\arguments{
\item{pcgr_tbl}{Parsed PCGR TSV tibble.}

\item{tiers}{PCGR tiers to keep. Default: 1, 2, 3, 4.}

\item{splice_vars}{Include non-coding splice region variants reported in PCGR?}
}
\value{
Character vector of genes.
}
\description{
Get Mutated Genes from PCGR
}
