% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\name{get_ensembl_db}
\alias{get_ensembl_db}
\title{Create AnnotationHub object to extract ensmbldb}
\usage{
get_ensembl_db(x)
}
\arguments{
\item{x}{AnnotationHub object}
}
\value{
Vector of gene and transcript ids
}
\description{
Create AnnotationHub object to extract ensmbldb
}
\examples{
x <- get_ensembl_db(x = ah)
}
