% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\name{get_refdata}
\alias{get_refdata}
\title{Get Reference Data File Paths}
\usage{
get_refdata(dataset)
}
\arguments{
\item{dataset}{Reference RNAsum dataset of interest.}
}
\value{
List with paths to internal and external reference datasets
(Counts, Targets and Name).
}
\description{
Get a list of paths to internal and external reference data, based on a
selected dataset name.
}
\examples{
x <- get_refdata(dataset = "TEST")
}
