% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\name{get_refgenes}
\alias{get_refgenes}
\title{Get Reference Genes}
\usage{
get_refgenes(p)
}
\arguments{
\item{p}{RNAsum params list.}
}
\value{
List of tibbles corresponding to cancer, oncokb, immune and hrd genes
}
\description{
Get a list of reference genes.
}
\examples{
p <- list(
  genes_cancer = system.file("extdata/genes/umccr_cancer_genes.2019-03-20.tsv.gz",
    package = "RNAsum.data"
  ),
  genes_oncokb = system.file(
    "extdata/OncoKB/CancerGenesList.txt.gz",
    package = "RNAsum.data"
  ),
  civic_var_summaries = system.file(
    "extdata/CIViC/01-Oct-2018-VariantSummaries.tsv.gz",
    package = "RNAsum.data"
  )
)
x <- get_refgenes(p)
}
