% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColours.R
\name{getColours}
\alias{getColours}
\title{Assign colours to different elements}
\usage{
getColours(elements)
}
\arguments{
\item{elements}{Elements of interest.}
}
\value{
Colour assignment to different elements
}
\description{
Assign colours to different elements
}
