% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glanceExprPlot.R
\name{glanceExprPlot}
\alias{glanceExprPlot}
\title{Generate boxplot presenting expression profiles for selected set of genes}
\usage{
glanceExprPlot(
  genes,
  data,
  targets,
  sampleName,
  int_cancer,
  ext_cancer,
  comp_cancer,
  add_cancer = NULL,
  hexcode,
  type = "z",
  sort = "diff",
  scaling = "gene-wise",
  report_dir
)
}
\arguments{
\item{genes}{Genes of interest.}

\item{data}{Input data.}

\item{targets}{Target groups.}

\item{sampleName}{Sample name.}

\item{int_cancer}{Internal cancer group.}

\item{ext_cancer}{External cancer group.}

\item{comp_cancer}{Complete cancer group.}

\item{add_cancer}{Addtional cancer type.}

\item{hexcode}{Hexcode.}

\item{type}{Type for expression values.}

\item{sort}{Sort order.}

\item{scaling}{Scaling type.}

\item{report_dir}{Reprot directory.}
}
\value{
Boxplot presenting expression profiles for selected set of genes.
}
\description{
Generates boxplot presenting expression profiles for selected set of genes.
}
