% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{immune_summary}
\alias{immune_summary}
\title{Get Immune Gene Summary}
\usage{
immune_summary(tbl_imarkers, tbl_igram = NULL, igram_param = TRUE)
}
\arguments{
\item{tbl_imarkers}{Tibble with immune response marker genes.}

\item{tbl_igram}{Tibble with immunogram genes.}

\item{igram_param}{Immunogram inclusion TRUE/FALSE parameter.}
}
\value{
Character vector of genes.
}
\description{
Get Immune Gene Summary
}
