% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{kallisto_counts}
\alias{kallisto_counts}
\title{Get Kallisto Counts}
\usage{
kallisto_counts(x, tx2gene = NULL)
}
\arguments{
\item{x}{Path to \code{abundance.tsv} file with abundances. See \link[tximport:tximport]{tximport::tximport}.}

\item{tx2gene}{data.frame with tx_name and gene_id columns. See \link[tximport:tximport]{tximport::tximport}.}
}
\value{
Tibble with the counts per gene transcript, or NULL if any of the
input params are NULL.
}
\description{
Get kallisto counts via tximport.
}
\examples{
x <- system.file("rawdata/test_data/quant/abundance.tsv", package = "RNAsum")
tx2gene <- NULL
(kc <- kallisto_counts(x, tx2gene)) # NULL since no tx2gene specified
}
