% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions.R
\name{known_trans}
\alias{known_trans}
\title{Get Known Translocations}
\usage{
known_trans(kt_fusiongdb, kt_cgi)
}
\arguments{
\item{kt_fusiongdb}{FusionGDB parsed tibble.}

\item{kt_cgi}{CGI parsed tibble.}
}
\value{
Tibble with known translocations.
}
\description{
Flag known fusions based on info from:
\itemize{
\item FusionGDB (https://ccsm.uth.edu/FusionGDB)
\item Cancer Biomarkers database (CGI) (https://www.cancergenomeinterpreter.org/biomarkers)
}
}
