% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions.R
\name{known_translocations_cgi_process}
\alias{known_translocations_cgi_process}
\title{Process CGI Known Translocations}
\usage{
known_translocations_cgi_process(kt_cgi)
}
\arguments{
\item{kt_cgi}{Parsed CGI tibble.}
}
\value{
Tibble with dups removed.
}
\description{
Processes parsed CGI known translocations tibble, removing duplicated
translocations that have a different source.
}
