% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutCNexprPlot.R
\name{mutCNexprPlot}
\alias{mutCNexprPlot}
\title{Generate scatterplot with per-gene expression values (y-axis), CN values (x-axis) and mutation status info (colours), if provided}
\usage{
mutCNexprPlot(
  data,
  alt_data = FALSE,
  cn_bottom = cn_bottom,
  cn_top = cn_top,
  comp_cancer,
  type = "z",
  report_dir
)
}
\arguments{
\item{data}{Input data.}

\item{alt_data}{Boolean (Generates scatterplot with per-gene expression values (y-axis)).}

\item{cn_bottom}{Bottom value for copy number.}

\item{cn_top}{Top value for copy number.}

\item{comp_cancer}{Complete cancer group.}

\item{type}{Type.}

\item{report_dir}{Report directory.}
}
\value{
Scatterplot with per-gene expression values (y-axis), CN values (x-axis) and mutation status info (colours), if provided.
}
\description{
Generates scatterplot with per-gene expression values (y-axis), CN values (x-axis) and mutation status info (colours), if provided
}
