% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_position.R
\name{nearest_position}
\alias{nearest_position}
\title{Check for nearest value in a vector}
\usage{
nearest_position(vector, x)
}
\arguments{
\item{vector}{Input vector}

\item{x}{Value to search}
}
\value{
Nearest value in a vector
}
\description{
Check for nearest value in a vector
}
