% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Perform PCA}
\usage{
pca(data, targets, title = "", report_dir, suffix = "")
}
\arguments{
\item{data}{Input data.}

\item{targets}{Target group.}

\item{title}{Title.}

\item{report_dir}{Report directory.}

\item{suffix}{Suffix for output file.}
}
\value{
Outputs a list with data frame and samples colouring info ready for plotting.
}
\description{
Perform PCA. Outputs a list with dataframe and samples colouring info ready for plotting.
}
