% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgr.R
\name{pcgr_expr}
\alias{pcgr_expr}
\title{Get PCGR genes with hits}
\usage{
pcgr_expr(pcgr = NULL, expr_data.z, expr_data.perc)
}
\arguments{
\item{pcgr}{Path to parsed PCGR tiered TSV tibble if available.}

\item{expr_data.z}{Tibble with gene symbol and z-score expr values.}

\item{expr_data.perc}{Tibble with gene symbol and percentile expr values.}
}
\value{
List.
}
\description{
Get PCGR genes with hits
}
