% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_rank.R
\name{perc_rank}
\alias{perc_rank}
\title{Convert a vector of numbers into corresponding vector of their percentiles}
\usage{
perc_rank(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
Vector of percentiles
}
\description{
Convert a vector of numbers into corresponding vector of their percentiles
}
