% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{plot_cnv_manhattan}
\alias{plot_cnv_manhattan}
\title{Manhattan Plot for Copynumber Data}
\usage{
plot_cnv_manhattan(
  cnv_data,
  genes_to_highlight = NULL,
  title = "",
  show_legend = TRUE,
  point_size = 3,
  highlight_size = 8,
  cn_top = NULL,
  cn_bottom = NULL
)
}
\arguments{
\item{cnv_data}{CN data frame including Gene, P, Diff_Perc, Diff_Z_score,
Alterations, GENEBIOTYPE, ENSEMBL, SEQNAME, GENESEQSTART, GENESEQEND,
Immunic_Cycle_Role.}

\item{genes_to_highlight}{Atomic vector of gene names to highlight.}

\item{title}{Plot title.}

\item{show_legend}{Show legend.}

\item{point_size}{Point size.}

\item{highlight_size}{Highlight size.}

\item{cn_top}{Top CN threshold.}

\item{cn_bottom}{Bottom CN threshold.}
}
\value{
Plotly object.
}
\description{
Manhattan Plot for Copynumber Data
}
\examples{
\dontrun{
cnv_data <- data.annot
genes_to_highlight <- unique(data.sub$Gene)
}
}
