% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{ppl_cnv_som_gene_read}
\alias{ppl_cnv_som_gene_read}
\title{Read PURPLE CNV Gene File}
\usage{
ppl_cnv_som_gene_read(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.gene.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{purple.cnv.gene.tsv} file, which summarises copy number
alterations of each gene in the HMF panel
(see https://github.com/hartwigmedical/hmftools/tree/master/purple#gene-copy-number-file).
}
