% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare2write.R
\name{prepare2write}
\alias{prepare2write}
\title{Prepare object to write into a file}
\usage{
prepare2write(x)
}
\arguments{
\item{x}{Input object.}
}
\value{
Object for writing to a file.
}
\description{
Prepare object to write into a file
}
