% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{purple_cnv_summary}
\alias{purple_cnv_summary}
\title{Get PURPLE CNV Summary}
\usage{
purple_cnv_summary(tbl, cancer_genes_symbol, cn_bottom, cn_top)
}
\arguments{
\item{tbl}{PURPLE gene CNV tibble}

\item{cancer_genes_symbol}{Character vector of gene symbols to filter on.}

\item{cn_bottom}{CN threshold value to classify genes within lost regions.}

\item{cn_top}{CN threshold value to classify genes within gained regions.}
}
\value{
List with genes, top and bottom copy number thresholds.
}
\description{
Get PURPLE CNV Summary
}
