% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions.R
\name{rcircos_cyto_info38}
\alias{rcircos_cyto_info38}
\title{RCircos HG38 CytoBandIdeogram Object}
\usage{
rcircos_cyto_info38()
}
\value{
A dataframe with hg38 cytoband/ideogram information from RCircos.
}
\description{
RCircos HG38 CytoBandIdeogram Object
}
