% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions.R
\name{rcircos_plot}
\alias{rcircos_plot}
\title{RCircos Plot}
\usage{
rcircos_plot(df_circos, df_circos_pairs, cyto.info)
}
\arguments{
\item{df_circos}{Dataframe with 4 columns: Chromosome, chromStart, chromEnd, and Gene.}

\item{df_circos_pairs}{Dataframe with 3 columns for each fusion mate:
Chromosome, chromStart and chromEnd for geneA and geneB.}

\item{cyto.info}{A dataframe with hg38 cytoband/ideogram information from RCircos.}
}
\value{
An RCircos plot object.
}
\description{
RCircos Plot
}
