% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read_sample_data}
\alias{read_sample_data}
\title{Read Sample Data}
\usage{
read_sample_data(p, results_dir, tx2gene = NULL)
}
\arguments{
\item{p}{RNAsum params list.}

\item{results_dir}{Directory to output extracted Arriba PNGs to (created
automatically if it does not already exist).}

\item{tx2gene}{data.frame with tx_name and gene_id columns. Required for salmon.
See \link[tximport:tximport]{tximport::tximport}.}
}
\value{
A list of the input sample data.
}
\description{
Reads sample data, including Arriba fusions, Arriba plots, Salmon counts,
DRAGEN fusions, DRAGEN mapping metrics, Manta SVs, PURPLE CNVs, and PCGR SNVs.
}
\examples{
p <- list(
  dragen_wts_dir = system.file("rawdata/test_data/dragen", package = "RNAsum"),
  arriba_dir = system.file("rawdata/test_data/dragen/arriba", package = "RNAsum"),
  umccrise = system.file(
    "rawdata/test_data/umccrised/test_sample_WGS",
    package = "RNAsum"
  )
)
res <- read_sample_data(p, tempdir())
}
