% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read_wgs_data}
\alias{read_wgs_data}
\title{Read WGS Data}
\usage{
read_wgs_data(p)
}
\arguments{
\item{p}{RNAsum params list.}
}
\value{
A list of the input sample data.
}
\description{
Reads WGS data, including PCGR \code{tiers.tsv}, PURPLE \code{cnv.gene.tsv}, and Manta
\code{manta.tsv}. If the file path has been specified in the RNAsum params and is
valid, it is returned. As a fallback, if the umccrise directory param has
been specified, then there is an attempt to detect the file pattern in there.
}
\examples{
p <- list(
  umccrise = system.file("rawdata/test_data/umccrised/test_sample_WGS", package = "RNAsum"),
  pcgr_tiers_tsv = system.file(
    "rawdata/test_data/umccrised/test_sample_WGS/small_variants",
    "TEST-somatic.pcgr.snvs_indels.tiers.tsv",
    package = "RNAsum"
  ),
  manta_tsv = system.file(
    "rawdata/test_data/umccrised/test_sample_WGS/structural/TEST-prioritize-manta.tsv",
    package = "RNAsum"
  )
)
(res <- read_wgs_data(p))
}
