% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{REFERENCE_DATASETS}
\alias{REFERENCE_DATASETS}
\title{RNAsum Reference Datasets}
\format{
A list of lists with the project abbreviation code, name,
tissue code and number of samples for each of the 40 datasets.
}
\usage{
data(REFERENCE_DATASETS)
}
\description{
Reference datasets available in RNAsum.
}
\keyword{datasets}
