% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shortcuts.R
\docType{package}
\name{RNAsum-package}
\alias{RNAsum}
\alias{RNAsum-package}
\title{RNAsum: A Robust Tool for Performing Transcriptome Profiling on Cancer Data}
\description{
RNA-seq reporting workflow designed to post-process, summarise and visualise an output from bcbio-nextgen RNA-seq or Dragen RNA pipelines. Its main application is to complement genome-based findings and provide additional evidence for detected alterations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://umccr.github.io/RNAsum/}
  \item \url{https://github.com/umccr/RNAsum}
}

}
\author{
\strong{Maintainer}: Sehrish Kanwal \email{kanwals@unimelb.edu.au} (\href{https://orcid.org/0000-0002-5044-4692}{ORCID})

Authors:
\itemize{
  \item Jacek Marzec \email{jacek.marzec@unimelb.edu.au}
}

}
\keyword{internal}
