% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{rnasum_rmd}
\alias{rnasum_rmd}
\title{Generate UMCCR RNAsum Report}
\usage{
rnasum_rmd(out_file = NULL, quiet = FALSE, pars)
}
\arguments{
\item{out_file}{Path to output HTML file (needs '.html' suffix).}

\item{quiet}{Suppress printing during rendering.}

\item{pars}{List of named parameters that override custom params specified within the Rmd's YAML front-matter.}
}
\value{
Path to rendered HTML report.
}
\description{
Generates a UMCCR RNAsum HTML report. It does so with the following steps:
\enumerate{
\item copy the rmd into tmp/rnasum.Rmd
\item render the rmd inside tmp/
\item return the path to the output HTML
}
}
