% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveWidgetFix.R
\name{saveWidgetFix}
\alias{saveWidgetFix}
\title{Wrapper for saveWidget bug fix}
\usage{
saveWidgetFix(widget, file, selfcontained = TRUE, ...)
}
\arguments{
\item{widget}{Widget to save.}

\item{file}{Full path to save HTML into.}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}

\item{...}{Additional arguments passed to \link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget}.}
}
\value{
Change work dir to compensate for saveWidget bug
}
\description{
A wrapper to saveWidget which compensates for bug where the temporary
directory with widget dependencies is not deleted - see
https://github.com/ramnathv/htmlwidgets/issues/296.
}
