% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{session_info_kable}
\alias{session_info_kable}
\title{Session Information Kable}
\usage{
session_info_kable(pkgs = NULL)
}
\arguments{
\item{pkgs}{Vector of R packages to display in the vignette. By default returns all.}
}
\value{
A list with two kables containing information about the platform and
the specified packages.
}
\description{
Session information kables for vignettes.
}
