% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{session_info_tbls}
\alias{session_info_tbls}
\title{Session Information Tables}
\usage{
session_info_tbls(pkgs = NULL)
}
\arguments{
\item{pkgs}{Vector of R packages to display in the vignette. By default returns all.}
}
\value{
A list with two tibbles containing information about the platform and
the specified packages.
}
\description{
Session information tables.
}
