% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_prioritize.R
\name{split_sv_field}
\alias{split_sv_field}
\title{Prioritize SV events}
\usage{
split_sv_field(.data, field, is_pct = F)
}
\arguments{
\item{sv_file}{Input structural variants files.}
}
\value{
Prioritized variants list.
}
\description{
Prioritize SV events
}
