% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sv_manta_summary}
\alias{sv_manta_summary}
\title{Get Manta SV Summary}
\usage{
sv_manta_summary(tbl)
}
\arguments{
\item{tbl}{Tibble with melted SVs from umccrise, containing 'Genes' column.}
}
\value{
Character vector of Genes.
}
\description{
Get Manta SV Summary
}
