% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_prioritize_short.R
\name{sv_prioritize_short}
\alias{sv_prioritize_short}
\title{Prioritize structural variants events}
\usage{
sv_prioritize_short(sv_file)
}
\arguments{
\item{sv_file}{Input file.}
}
\value{
Prioritized variant calls.
}
\description{
Prioritize structural variants events
}
