% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{sv_process}
\alias{sv_process}
\title{Process SV Object}
\usage{
sv_process(sv_tsv_obj)
}
\arguments{
\item{sv_tsv_obj}{SV list object read via \code{sv_prioritize}.}
}
\value{
List with:
\itemize{
\item total SV (unmelted) variants
\item tibble with melted variants
\item genes involved in multi-gene events
}
}
\description{
Process SV Object
}
