% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sv_summary}
\alias{sv_summary}
\title{Get SV Summary}
\usage{
sv_summary(tbl)
}
\arguments{
\item{tbl}{Tibble with melted SVs from sash, containing 'Genes' column.}
}
\value{
Character vector of Genes.
}
\description{
Get SV Summary
}
