% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webplot.R
\name{webplot}
\alias{webplot}
\title{Generate spider web plots to present immunogram genes}
\usage{
webplot(
  data,
  data.row = NULL,
  y.cols = NULL,
  main = NULL,
  add = F,
  col = "red",
  lty = 1,
  scale = T
)
}
\arguments{
\item{data}{Dataframe or matrix.}

\item{data.row}{Row of data to plot (if NULL uses row 1).}

\item{y.cols}{Columns of interest (if NULL it selects all numeric columns).}

\item{main}{Title of plot (if NULL then rowname of data).}

\item{add}{Whether the plot should be added to an existing plot.}

\item{col}{Color of the data line}

\item{lty}{Lty of the data line}

\item{scale}{Scale.}
}
\value{
Spider web plots to present immunogram genes
}
\description{
Generate spider web plots to present immunogram genes
}
