% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allow_thumbnails.R
\name{allow_thumbnails}
\alias{allow_thumbnails}
\title{Generate a full-resolution pdf image before generating a smaller image}
\usage{
allow_thumbnails(x, options)
}
\arguments{
\item{x}{A pdf image.}

\item{options}{Additional settings.}
}
\value{
A full resolution pdf image.
}
\description{
Generates a full-resolution pdf image before generating a smaller image
in the chunk (from https://stackoverflow.com/questions/37834053/what-is-a-simple-way-to-thumbnail-some-plots-in-r-markdown-knitr).
}
