% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusions_arriba.R
\name{arriba_tsv_read}
\alias{arriba_tsv_read}
\title{Read Arriba TSV File}
\usage{
arriba_tsv_read(x = NULL)
}
\arguments{
\item{x}{Path to Arriba TSV file.}
}
\value{
A tibble with the contents of the input TSV file, or NULL if input is NULL or TSV has no data rows.
}
\description{
Reads the TSV file output by Arriba.
}
\examples{
x <- system.file("rawdata/test_data/dragen/arriba/fusions.tsv", package = "RNAsum")
(a <- arriba_tsv_read(x))
}
