% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDatasets.R
\name{combineDatasets}
\alias{combineDatasets}
\title{Combine sample expression profile with reference datasets.}
\usage{
combineDatasets(sample_name, sample_counts, ref_data, report_dir, dataset)
}
\arguments{
\item{sample_name}{Sample name.}

\item{sample_counts}{Sample counts.}

\item{ref_data}{Reference data.}

\item{report_dir}{Report directory.}

\item{dataset}{Sample expression profile.}
}
\value{
Combined sample expression profile with reference datasets.
}
\description{
Combine sample expression profile with reference datasets.
}
