% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen_fusions.R
\name{dragen_fusions_read}
\alias{dragen_fusions_read}
\title{Read DRAGEN Fusions}
\usage{
dragen_fusions_read(x)
}
\arguments{
\item{x}{Path to the \code{fusion_candidates.final} file.}
}
\value{
A tibble with the contents of the input TSV file.
}
\description{
Reads the \code{fusion_candidates.final} file output by DRAGEN.
}
\examples{
x <- system.file("rawdata/test_data/dragen",
                 "test_sample_WTS.fusion_candidates.final", package = "RNAsum")
(a <- dragen_fusions_read(x))
}
