% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\name{get_refgenes}
\alias{get_refgenes}
\title{Get Reference Genes}
\usage{
get_refgenes(p)
}
\arguments{
\item{p}{RNAsum params list.}
}
\value{
List of tibbles corresponding to cancer, oncokb, immune and hrd genes
}
\description{
Get a list of reference genes.
}
\examples{
p <- list(
  genes_cancer = system.file("rawdata/genes/umccr_cancer_genes.2019-03-20.tsv", package = "RNAsum"),
  genes_oncokb = system.file("rawdata/OncoKB/CancerGenesList.txt", package = "RNAsum"),
  genes_immune_markers = system.file("rawdata/genes/Genes_immune_markers.txt", package = "RNAsum"),
  genes_hrd = system.file("rawdata/genes/Genes_HRD.txt", package = "RNAsum")
)
x <- get_refgenes(p)
}
