% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read_sample_data}
\alias{read_sample_data}
\title{Read Sample Data}
\usage{
read_sample_data(p, results_dir, tx2gene = NULL)
}
\arguments{
\item{p}{RNAsum params list.}

\item{results_dir}{Directory to output extracted Arriba PNGs to (created
automatically if it does not already exist).}

\item{tx2gene}{data.frame with tx_name and gene_id columns. Required for salmon.
See \link[tximport:tximport]{tximport::tximport}.}
}
\value{
A list of the input sample data.
}
\description{
Reads sample data, including Arriba fusions, Arriba plots, Salmon counts, and
DRAGEN fusions.
}
\examples{
p <- list(
  dragen_rnaseq = system.file("rawdata/test_data/dragen", package = "RNAsum"),
  arriba_pdf = system.file("rawdata/test_data/dragen/arriba/fusions.pdf", package = "RNAsum"),
  arriba_tsv = system.file("rawdata/test_data/dragen/arriba/fusions.tsv", package = "RNAsum"),
  salmon = NULL,
  dragen_fusions = system.file(
    "rawdata/test_data/dragen/test_sample_WTS.fusion_candidates.final",
    package = "RNAsum"
  )
)
res <- read_sample_data(p, tempdir())
}
