% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{salmon_counts}
\alias{salmon_counts}
\title{Get Salmon Counts}
\usage{
salmon_counts(x, tx2gene)
}
\arguments{
\item{x}{Path to \code{quant.sf} file with abundances. See \link[tximport:tximport]{tximport::tximport}.}

\item{tx2gene}{See \link[tximport:tximport]{tximport::tximport}}
}
\value{
Tibble with the counts per gene transcript.
}
\description{
Returns salmon counts.
}
