% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveWidgetFix.R
\name{saveWidgetFix}
\alias{saveWidgetFix}
\title{Wrapper for saveWidget bug fix}
\usage{
saveWidgetFix(widget, file, ...)
}
\arguments{
\item{widget}{Input widget}

\item{file}{Input file}

\item{...}{}
}
\value{
Change work dir to compensate for saveWidget bug
}
\description{
A wrapper to saveWidget which compensates for arguable BUG in saveWidget which requires
\code{file} to be in current working directory (see https://github.com/ramnathv/htmlwidgets/issues/299).
}
