% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveWidgetFix.R
\name{saveWidgetFix}
\alias{saveWidgetFix}
\title{Wrapper for saveWidget bug fix}
\usage{
saveWidgetFix(widget, file, ...)
}
\arguments{
\item{widget}{Input widget}

\item{file}{Input file}

\item{...}{}
}
\value{
Change work dir to compensate for saveWidget bug
}
\description{
Wrapper for saveWidget bug fix
}
