% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_prioritize.R
\name{sv_prioritize}
\alias{sv_prioritize}
\title{Prioritize SV events}
\usage{
sv_prioritize(sv_file)
}
\arguments{
\item{sv_file}{Input structural variants files.}
}
\value{
Prioritized variants list.
}
\description{
Code from umccrise to prioritise SV events (version for "-manta.tsv" files)
}
